methods/spider-agent-lite/pyproject.toml
[project]
name = "spider-agent-lite"
version = "0.1.0"
description = "An Agent Method Baseline for Spider 2.0-Lite based on Docker environment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "tqdm>=4.66.0",
    "openai>=1.0.0",
    "docker>=7.0.0",
    "gymnasium>=0.29.0",
    "requests>=2.31.0",
    "tiktoken>=0.5.0",
    "dashscope>=1.14.0",
    "groq>=0.4.0",
    "google-generativeai>=0.3.0",
    "gdown>=4.7.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["spider_agent"]

